Do not error when user passes value under the max to setsockopt
This change makes it so that the setsockopt() function does not error
when a user passes a value for SO_RCVBUF equal or under the maximim
(kern.ipc.maxsockbuf).

Currently the behavior is to error if the value is greater than the
_adjusted_ max, which on amd64 turns out to be something like
kern.ipc.maxsockbuf * 0.889

This is confusing for users, and this change will set the value passed
to the adjusted max if the value is greater than the adjusted and less
than the actual max.